Pythonization: dead code elimination, modulo operator.
When changing pixel movement units, new value is calculated based on modulo of number of settings available.
In oe instance, the else statmeent did nothing and was dead code, and this is now gone.
When announcing pixel movement commands, rather than holding the string in a dedicated variable, just let it pass through ui.message.