More pythonization, code simplification, new position storage format. re #4.
Along with the new positions storage format (identical in structore to an ini file), more code refinements were done. Specifically, the new file format can be opened with ini parsers and dictionary transofmers, inclding ConfigObj (the module used in this add-on).

Plus, more refinements and pythonization work:
* When jumping to a mouse position, str.isdigit is no longer consulted. Instead, a direct integer casting and exception handling will be used, making the code simple.
Due to the new positions format, codecs and opening/reading/writing files directly is no longer supported. A migration tool will be done next.
Due to removal of codecs, in positions list dialog, positions wil be handled via ConfiObj and treated as a dictionary.
The new location for positions database is addonpath/savedPositions. Due to format changes, verion 1 and 2 config are not compatible.