Saved positions: completely redesigned GUI.
Previously, saved positions dialog used split layout where tags were displayed on the left side while the right side contained buttons. This isn't actually the way majority of dialogs are designed, in that many use vertical layout with tags on top and buttons at the bottom. To fill in the empty space, changes will be made to saved positions list box by adding coordinate columns.
Added close button to mimic that of NVDA's Add-ons Manager, and edited the dialog code to align with rest of NVDA.
Instead of OK and Cancel buttons, the new Jump button functions as OK button to jump to the selected tag (if possible).